Update for latest version of rails
Hello, i think this would be a major version bump but wanted to at least put it our there for discussion.  I installed guard-test on a rails 6 project and the options in the runner are no longer supported with newer versions of guard.   Using --color allowed to get the visual feedback and guard ran properly